Refactor assets view

+++ assets/forms.py
@@ -2,4 +2,4 @@ from django import forms
 
 
 class AssetCreateUpdateForm(forms.Form):
-    amount = forms.DecimalField()
+    amount = forms.FloatField()
